SubSync: Initial commit
First commit. I'm pretty terrible at Python, it's not something
I use very often - don't judge ;)

This commit allows for the getting of subscriptions from two
authenticated YouTube channels. It then provides a visual diff
of the subscriptions between the channels and allows the user
to select the channels they wish to subscribe to or skip.